Make dialog composible
